Use uint8 type for decimals

+++ contracts/token/ERC20/ERC20Custom.sol
@@ -38,7 +38,7 @@ contract ERC20Custom is Context, IERC20, IERC20Metadata {
     mapping(address => mapping(address => uint256)) private _allowances;
 
     uint256 private _totalSupply;
-    uint256 private _decimals;
+    uint8 private _decimals;
 
     string private _name;
     string private _symbol;
@@ -52,7 +52,7 @@ contract ERC20Custom is Context, IERC20, IERC20Metadata {
      * All two of these values are immutable: they can only be set once during
      * construction.
      */
-    constructor(string memory name_, string memory symbol_, uint256 decimals_) {
+    constructor(string memory name_, string memory symbol_, uint8 decimals_) {
         _name = name_;
         _symbol = symbol_;
         _decimals = decimals_;
